move serial driver registering out of device.cpp
